Conflicts:
	buttleofx/MainWindow.qml

+++ buttleofx/gui/paramEditor/wrappers/double2DWrapper.py
@@ -68,7 +68,7 @@ class Double2DWrapper(QtCore.QObject):
 
     paramType = QtCore.Property(unicode, getParamType, setParamType, notify=changed)
     text = QtCore.Property(unicode, getText, setText, notify=changed)
-    value1 = QtCore.Property(int, getValue1, setValue1, notify=changed)
-    value2 = QtCore.Property(int, getValue2, setValue2, notify=changed)
+    value1 = QtCore.Property(float, getValue1, setValue1, notify=changed)
+    value2 = QtCore.Property(float, getValue2, setValue2, notify=changed)
     maximum = QtCore.Property(float, getMaximum, setMaximum, notify=changed)
     minimum = QtCore.Property(float, getMinimum, setMinimum, notify=changed)
